+++ src/events/messageUpdate.js
@@ -1,6 +1,8 @@
 const messageUpdate = (client, oldMessage, newMessage) => {
-  if (newMessage.channel.name.startsWith('count-to-')) {
-    newMessage.delete();
+  if (client.features.countToChannel) {
+    if (newMessage.channel.name.startsWith('count-to-')) {
+      newMessage.delete();
+    }
   }
 };
